flavors-server/chat_project/requirements.txt
asgiref==3.8.1
channels==4.2.0
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.1
